requirements.txt
nonebot2[fastapi]~=2.4.2
nonebot-adapter-onebot~=2.2.3
nonebot-plugin-apscheduler~=0.2.0
pydantic~=1.10.13
python-dotenv~=1.0.1